server/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
numpy
mlx
mlx-lm
requests
pillow
python-dotenv
rich
python-json-logger
huggingface-hub
structlog
coremltools
